+++ napari/components/experimental/__init__.py
@@ -1,1 +1,2 @@
 from .chunk import chunk_loader
+from .monitor.service import monitor
